- Displaying highlights when a page is loaded - Implement JWT tokens for authorization

+++ highlighter-backend/highlighter/routes/highlights.js
@@ -3,6 +3,7 @@ var debug = require('debug')('highlighter:highlights');
 var router = express.Router();
 const Joi = require('joi');
 const async = require('async');
+const auth = require('./auth.js')
 
 //Imported models
 const Highlights = require('../server/models').highlights;
@@ -27,7 +28,7 @@ let successResponse = {
 }
 
 //Process POST request of adding new highlight for user
-router.post('/new', (req, res, next) => {
+router.post('/new', auth.authenticateJWT,(req, res, next) => {
     debug("Req.body: ", req.body);
   const schema = Joi.object().keys({
     userid: Joi.number().required().label('Userid'),
@@ -43,6 +44,10 @@ router.post('/new', (req, res, next) => {
     return res.status(400).send(
       badRequest
     );
+  }
+
+  if (req.body.userid != req.user.id) {
+    return res.status(403).send(auth.unauthorizedRes);
   }
 
   let createJson = req.body;
@@ -62,48 +67,79 @@ router.post('/new', (req, res, next) => {
   });
 });
 
-router.get('/', (req, res, next) => {
+router.get('/',  auth.authenticateJWT, (req, res, next) => {
   debug("Req.body: ", req.query);
-const schema = Joi.object().keys({
-  userid: Joi.number().required().label('Userid'),
-  size: Joi.number().label("Number of highlights"),
-  page: Joi.number().required().label("Page number"),
-  to_include: Joi.number().default(0).label("Number of records to include"),
-});
-
-const joiResult = Joi.validate(req.query, schema);
-if (joiResult.error) {
-  badRequest.data = joiResult.error;
-  return res.status(400).send(
-    badRequest
-  );
-}
-let size = req.query.size ? req.query.size : 10;
-let to_include = req.query.to_include == undefined ? 0 : req.query.to_include;
-Highlights
-  .findAll({ 
-    where: { userid: req.query.userid }, 
-    limit: size,
-    offset: ((req.query.page - 1) * size) - to_include})
-  .then(highlight => {
-    debug('Success: ', highlight.length);
-    successResponse.data = highlight;
-    return res.status(200).send(
-      successResponse
-    );
-  })
-  .catch(error => {
-    debug('Error: ', error)
-    serverError.data = error;
-    return res.status(500).send(
-      serverError
-    );
+  const schema = Joi.object().keys({
+    userid: Joi.number().required().label('Userid'),
+    type: Joi.string().required().label('Type of query'),
+    size: Joi.number().label("Number of highlights"),
+    page: Joi.number().label("Page number"),
+    to_include: Joi.number().default(0).label("Number of records to include"),
+    url: Joi.string().default('').label("URL"),
   });
+
+  const joiResult = Joi.validate(req.query, schema);
+  if (joiResult.error) {
+    badRequest.data = joiResult.error;
+    return res.status(400).send(
+      badRequest
+    );
+  }
+
+  if (req.query.userid != req.user.id) {
+    return res.status(403).send(auth.unauthorizedRes);
+  }
+
+  if (req.query.type === 'popup') {
+    let size = req.query.size ? req.query.size : 10;
+    let to_include = req.query.to_include == undefined ? 0 : req.query.to_include;
+    Highlights
+    .findAll({ 
+      where: { userid: req.query.userid }, 
+      limit: size,
+      offset: ((req.query.page - 1) * size) - to_include,
+      url: req.query.url
+    })
+    .then(highlight => {
+      debug('Success: ', highlight.length);
+      successResponse.data = highlight;
+      return res.status(200).send(
+        successResponse
+      );
+    })
+    .catch(error => {
+      debug('Error: ', error)
+      serverError.data = error;
+      return res.status(500).send(
+        serverError
+      );
+    });
+  } else if (req.query.type === 'content') {
+    Highlights
+    .findAll({ 
+      where: { userid: req.query.userid, url: req.query.url }
+    })
+    .then(highlight => {
+      debug('Success: ', highlight.length);
+      successResponse.data = highlight;
+      return res.status(200).send(
+        successResponse
+      );
+    })
+    .catch(error => {
+      debug('Error: ', error)
+      serverError.data = error;
+      return res.status(500).send(
+        serverError
+      );
+    });
+  }
 });
 
-router.delete('/', (req, res, next) => {
+router.delete('/', auth.authenticateJWT, (req, res, next) => {
   debug("Req.body: ", req.body);
   const schema = Joi.object().keys({
+    userid: Joi.number().required().label('Userid'),
     highlighterid: Joi.number().required().label('Highlighter ID'),
   });
 
@@ -114,9 +150,14 @@ router.delete('/', (req, res, next) => {
     badRequest
   );
   }
+
+  if (req.body.userid != req.user.id) {
+    return res.status(403).send(auth.unauthorizedRes);
+  }
+  
   Highlights
   .destroy({ 
-    where: { id: req.body.highlighterid }})
+    where: { id: req.body.highlighterid, userid: req.body.userid }})
   .then(numberOfDeletes => {
     successResponse.message = 'Highlight deleted successfully!'
     successResponse.data = numberOfDeletes;
